feat(gatsby): Support incremental builds with collection routes
fix lint
