+++ packages/gatsby-plugin-page-creator/.gitignore
@@ -29,4 +29,3 @@ node_modules
 decls
 dist
 
-/*.js
