+++ packages/gatsby-plugin-page-creator/gatsby-node.js
@@ -1,0 +1,5 @@
+exports.createPagesStatefully = require("./dist/lifecycles/create-pages-statefully").createPagesStatefully
+exports.createPages = require("./dist/lifecycles/create-pages").createPages
+exports.onPreInit = require("./dist/lifecycles/on-pre-init").onPreInit
+exports.onDeleteNode = require("./dist/lifecycles/on-delete-node").onDeleteNode
+exports.setFieldsOnGraphQLNodeType = require("./dist/lifecycles/set-fields-on-graphql-node-type").setFieldsOnGraphQLNodeType
