+++ packages/gatsby-plugin-page-creator/src/__tests__/path-utils.ts
@@ -1,0 +1,11 @@
+import { extractModel } from "../path-utils"
+
+describe(`Path Utils`, () => {
+  it(`extractModel`, () => {
+    expect(
+      extractModel(
+        `/Users/blainekasten/Sites/unified-routes/src/pages/assets/{DatoCmsWork.slug}.js`
+      )
+    ).toBe(`DatoCmsWork`)
+  })
+})
