+++ packages/gatsby-plugin-page-creator/src/collection-routes/is-valid-collection-path-implementation.ts
