+++ packages/gatsby-plugin-page-creator/src/known-collections.ts
@@ -1,0 +1,2 @@
+// This should be stored in state somehow I think
+export default new Map()
