+++ packages/gatsby-plugin-page-creator/src/types.ts
@@ -1,0 +1,7 @@
+import { PluginOptions } from "gatsby"
+
+export interface IOptions extends PluginOptions {
+  path: string
+  pathCheck: boolean
+  ignore: Array<string>
+}
