+++ packages/gatsby/scripts/__tests__/api.js
@@ -46,6 +46,7 @@ it("generates the expected api output", done => {
           "onCreateBabelConfig": Object {},
           "onCreateDevServer": Object {},
           "onCreateNode": Object {},
+          "onDeleteNode": Object {},
           "onCreatePage": Object {},
           "onCreateWebpackConfig": Object {},
           "onPostBootstrap": Object {},
